refactor(markdown): split markdown utils into separate files
Move each util from utils.ts into its own file under utils/, one
utility per file, and re-export them from index.ts so the public API
of @tiptap/markdown stays unchanged.
