+++ packages/markdown/src/index.ts
@@ -1,3 +1,11 @@
 export * from './Extension.js'
 export * from './MarkdownManager.js'
-export * from './utils.js'
+export * from './utils/assumeContentType.js'
+export * from './utils/closeMarksBeforeNode.js'
+export * from './utils/extractAbsorbedBlankLines.js'
+export * from './utils/findMarksToClose.js'
+export * from './utils/findMarksToCloseAtEnd.js'
+export * from './utils/findMarksToOpen.js'
+export * from './utils/isTaskItem.js'
+export * from './utils/reopenMarksAfterNode.js'
+export * from './utils/wrapInMarkdownBlock.js'
